- Celebrating our 200th revision with a haiku in Settings dialog. - Settings::ApplyChanges() is now functional. - Auto-synchronization now works at application start-up. - Double-clicking on items in History dialog now opens Anime Information dialog. - Added an option to not get notified when an episode is not recognized. - Added an explanation text for root folders to Settings dialog. - Added an explanation text for torrent filters to Settings dialog. - Changed default settings confirmation dialog's icon to a warning sign, since the process involves some risk. - Simplified presentation of the last updated value of season data. - Rounded up all HttpClients into a single structure. - Removed unnecessary HttpClientMode values. - Renamed a couple of things around the Anime Information dialog.
git-svn-id: http://taiga.googlecode.com/svn/trunk@200 77add795-4bdf-f0f6-a574-f4d068491005

+++ theme.cpp
@@ -106,12 +106,12 @@ bool Theme::LoadImages() {
   
   // Populate image lists
   HBITMAP hBitmap;
-  for (int i = 0; i < ICONCOUNT_16PX && i < icons16_.size(); i++) {
+  for (size_t i = 0; i < ICONCOUNT_16PX && i < icons16_.size(); i++) {
     hBitmap = GdiPlus.LoadImage(folder_ + L"16px\\" + icons16_.at(i) + L".png");
     ImgList16.AddBitmap(hBitmap, CLR_NONE);
     DeleteObject(hBitmap);
   }
-  for (int i = 0; i < ICONCOUNT_24PX && i < icons24_.size(); i++) {
+  for (size_t i = 0; i < ICONCOUNT_24PX && i < icons24_.size(); i++) {
     hBitmap = GdiPlus.LoadImage(folder_ + L"24px\\" + icons24_.at(i) + L".png");
     ImgList24.AddBitmap(hBitmap, CLR_NONE);
     DeleteObject(hBitmap);
